Karthik/Pat - Change package structure to move new activities out of refactor package to activity package
