+++ app/src/main/java/org/mercycorps/translationcards/activity/addTranslation/EnterTranslatedPhraseActivity.java
@@ -1,4 +1,4 @@
-package org.mercycorps.translationcards.refactor.activity;
+package org.mercycorps.translationcards.activity.addTranslation;
 
 import android.widget.TextView;
