+++ app/src/main/java/org/mercycorps/translationcards/activity/addTranslation/GetStartedActivity.java
@@ -1,8 +1,5 @@
-package org.mercycorps.translationcards.refactor.activity;
+package org.mercycorps.translationcards.activity.addTranslation;
 
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.widget.ImageButton;
 import android.widget.TextView;
 
 import org.mercycorps.translationcards.R;
