--- app/src/main/java/org/mercycorps/translationcards/data/TranslationFactory.java
@@ -1,8 +1,0 @@
-package org.mercycorps.translationcards.data;
-
-public class TranslationFactory {
-
-    public static Translation create() {
-        return new Translation();
-    }
-}
