+++ app/src/main/java/org/mercycorps/translationcards/exception/RecordAudioException.java
@@ -1,4 +1,4 @@
-package org.mercycorps.translationcards.refactor.activity;
+package org.mercycorps.translationcards.exception;
 
 public class RecordAudioException extends Exception {
