+++ app/src/test/java/org/mercycorps/translationcards/activity/TranslationsActivityTest.java
@@ -18,8 +18,7 @@ import org.mercycorps.translationcards.data.DbManager;
 import org.mercycorps.translationcards.data.Deck;
 import org.mercycorps.translationcards.data.Dictionary;
 import org.mercycorps.translationcards.data.Translation;
-import org.mercycorps.translationcards.refactor.activity.AddTranslationActivity;
-import org.mercycorps.translationcards.refactor.activity.GetStartedActivity;
+import org.mercycorps.translationcards.activity.addTranslation.GetStartedActivity;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricGradleTestRunner;
 import org.robolectric.RuntimeEnvironment;
