--- app/src/test/java/org/mercycorps/translationcards/refactor/activity/DictionaryFactory.java
@@ -1,10 +1,0 @@
-package org.mercycorps.translationcards.refactor.activity;
-
-import org.mercycorps.translationcards.data.Dictionary;
-
-public class DictionaryFactory {
-    public static Dictionary createDictionary(String name, int translationCount){
-        Dictionary dict = new Dictionary(name);
-        return dict;
-    }
-}
